requirements.txt
# Core Framework
streamlit==1.28.1
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1

# LLM & Embeddings
langchain>=0.1.0
sentence-transformers==3.0.0
ollama==0.1.0

# Vector Databases
faiss-cpu==1.8.0
chromadb>=0.4.0

# LLM API Clients
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Utilities
requests==2.31.0

PyYAML>=6.0.3